projects/echo_forge/watcher/requirements.txt
watchdog>=2.1.9
pytest>=7.0.0
PyYAML>=6.0
